destructure test01
