+++ packages/main/public/index.html
@@ -21,6 +21,5 @@
     <main id="main"></main>
     <footer class="footer1"></footer>
     <!-- End Footer Area -->
-    
 </body>
 </html>
