+++ packages/main/src/bootstrap.js
@@ -69,7 +69,7 @@ function userMain() {
     });
 
     mobileExe( mainEle, footerEle );
-
+    
 }
 
 // Loading Guest Page
